requirements.txt
gymnasium
numpy
pygame
PyOpenGL
PyOpenGL_accelerate
matplotlib
imageio
stable-baselines3
torch
tensorboard
opencv-python
